+++ database/migrations/2020_07_12_164752_add_field_rate_to_writers.php
@@ -14,7 +14,7 @@ class AddFieldRateToWriters extends Migration
     public function up()
     {
         Schema::table('writers', function (Blueprint $table) {
-            $table->integer('rate');
+            $table->double('rate', 3,1);
         });
     }
