+++ database/migrations/2020_07_12_164810_add_field_rate_to_pub_companies.php
@@ -14,7 +14,7 @@ class AddFieldRateToPubCompanies extends Migration
     public function up()
     {
         Schema::table('pub_companies', function (Blueprint $table) {
-            $table->integer('rate');
+            $table->double('rate', 3,1);
         });
     }
